관리자 아이디로 암호 초기화 하기

+++ src/main/java/com/choong/spr/controller/MemberController.java
@@ -1,6 +1,9 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
 import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -91,43 +94,73 @@ public class MemberController {
 	}
 	
 	@GetMapping("get")
-	public void getMember(String id, Model model) {
-		MemberDto member = service.getMemberById(id);
+	public String getMember(String id, Model model, Principal principal, HttpServletRequest request) {
+		if (hasAuthOrAdmin(id, principal, request)) {
+			
+			MemberDto member = service.getMemberById(id);
+			model.addAttribute("member", member);		
+			
+			return null; // return값 null로 지정시 views 폴더로 경로 잡힘
+		} 
 		
-		model.addAttribute("member", member);		
+		return "redirect:/member/login";
+	}
+	
+	private boolean hasAuthOrAdmin(String id, Principal principal, HttpServletRequest req) {
+		return req.isUserInRole("ROLE_ADMIN") || (principal != null && principal.getName().equals(id));
 	}
 	
 	@PostMapping("remove")
-	public String removeMember(MemberDto dto, RedirectAttributes rttr) {
-		boolean success = service.removeMember(dto);
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "회원 탈퇴 되었습니다.");
-			return "redirect:/board/list";
+	public String removeMember(MemberDto dto, Principal principal, HttpServletRequest req, RedirectAttributes rttr) {
+		if (hasAuthOrAdmin(dto.getId(), principal, req)) {
+			boolean success = service.removeMember(dto);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "회원 탈퇴 되었습니다.");
+				return "redirect:/board/list";
+			} else {
+				rttr.addAttribute("id", dto.getId());
+				return "redirect:/member/get";
+			}			
 		} else {
-			rttr.addAttribute("id", dto.getId());
 			return "redirect:/member/get";
 		}
 	}
 	
 	@PostMapping("modify")
-	public String modifyMember(MemberDto dto, String oldPassword, RedirectAttributes rttr) {
-		boolean success = service.modifyMember(dto, oldPassword);
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+	public String modifyMember(MemberDto dto, String oldPassword, Principal principal, HttpServletRequest req, RedirectAttributes rttr) {
+		if (hasAuthOrAdmin(dto.getId(), principal, req)) {
+			boolean success = service.modifyMember(dto, oldPassword);
+			
+			if (success) {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
+			}
+			
+			rttr.addFlashAttribute("member", dto); // model object
+			rttr.addAttribute("id", dto.getId()); // query string
+			
+			return "redirect:/member/get";			
 		} else {
-			rttr.addFlashAttribute("message", "회원 정보가 수정되지 않았습니다.");
-		}
-		
-		rttr.addFlashAttribute("member", dto); // model object
-		rttr.addAttribute("id", dto.getId()); // query string
-		
-		return "redirect:/member/get";
+			return "redirect:/member/login";
+		}		
 	}
 	
 	@GetMapping("login")
 	public void loginPage() {
 		
+	}
+	
+	@GetMapping("initpw")
+	public void initpwPage() {
+		
+	}
+	
+	@PostMapping("initpw")
+	public String initpwProcess(String id) {
+		service.initPassword(id);
+		
+		return "redirect:/board/list";
 	}
 }
